Load descendent node IDs for topic nodes
As with calculating the total size for topic nodes, we can get a list
of descendent nodes by brute force. This is useful to quickly find if
a content node has children which are in the collection. We can use that
information to produce a helpful indeterminate state for checkboxes when
selecting new content nodes.

https://phabricator.endlessm.com/T34350