Use a separate view to add content for a channel
Instead of CollectionEditorChannelPage displaying a list of all channel
content up front, we will move that to a different view that works like
a modal dialog. With this change, it will be easier to see up front
which content has been selected for a channel. It will also be easier to
add functionality which applies specifically to selected content.

https://phabricator.endlessm.com/T34712