Use a separate view to add content for a channel
Instead of CollectionEditorChannelPage displaying a list of all channel
content up front, we will move that to a different view which works like
a modal dialog. With this change, it will be easier to see which content
has been selected for a channel. In addition, we can use this new view
to add functionality which applies specifically to selected content.

https://phabricator.endlessm.com/T34712